# Conflicts:
#	src/main/java/com/cg/dto/HotelResponseDTO.java
#	src/main/java/com/cg/entity/Reservation.java
#	src/main/java/com/cg/entity/RoomType.java
#	src/main/java/com/cg/service/AmenityService.java
#	src/main/java/com/cg/service/HotelService.java
#	src/main/java/com/cg/service/PaymentService.java

+++ src/main/java/com/cg/entity/Amenity.java
@@ -7,16 +7,16 @@ import jakarta.persistence.*;
 public class Amenity {
 
     @Id
-    private Integer amenityId;
+    private Long amenityId;
 
     private String name;
     private String description;
 
 
-    public Integer getAmenityId() {
+    public Long getAmenityId() {
 		return amenityId;
 	}
-	public void setAmenityId(Integer amenityId) {
+	public void setAmenityId(Long amenityId) {
 		this.amenityId = amenityId;
 	}
 	public String getName() {
